fix(ci): fix all CI failures
- Fix feature-service clippy warning (unnecessary u64 cast)
- Move audit.toml to .cargo/audit.toml for cargo-audit
- Update CI workflow to- Update Cargo.toml workspace metadata

All checks pass locally:
- cargo clippy --workspace -- -D warnings
- cargo audit --deny warnings

+++ crates/feature-service/src/service.rs
@@ -272,7 +272,7 @@ impl FeatureService {
             // Use tenant_id as the hash key for consistent rollout
             let hash = Self::hash_tenant_id(tenant_id);
             #[allow(clippy::cast_possible_truncation)]
-            let threshold = (feature.rollout_percentage as u64 * u64::MAX / 100) as u64;
+            let threshold = feature.rollout_percentage as u64 * u64::MAX / 100;
 
             if hash > threshold {
                 return Ok(FeatureCheckResult {
